+++ client/common/admin/firewall/elements/index.ts
@@ -1,0 +1,5 @@
+// SPDX-FileCopyrightText: 2024 John Livingston <https://www.john-livingston.fr/>
+//
+// SPDX-License-Identifier: AGPL-3.0-only
+
+import './admin-firewall'