+++ server/lib/routers/api.ts
@@ -14,6 +14,7 @@ import { initFederationServerInfosApiRouter } from './api/federation-server-info
 import { initConfigurationApiRouter } from './api/configuration'
 import { initPromoteApiRouter } from './api/promote'
 import { initEmojisRouter } from './emojis'
+import { initAdminFirewallApiRouter } from './api/admin/firewall'
 
 /**
  * Initiate API routes
@@ -44,6 +45,8 @@ async function initApiRouter (options: RegisterServerOptions): Promise<Router> {
   await initConfigurationApiRouter(options, router)
   await initPromoteApiRouter(options, router)
   await initEmojisRouter(options, router)
+
+  await initAdminFirewallApiRouter(options, router)
 
   if (isDebugMode(options)) {
     // Only add this route if the debug mode is enabled at time of the server launch.